+++ test/boruta/oauth/integration/preauthorize_test.exs
@@ -68,8 +68,8 @@ defmodule Boruta.OauthTest.PreauthorizeTest do
                   error: :invalid_client,
                   error_description: "Invalid client_id or redirect_uri.",
                   status: :unauthorized,
-                  format: :fragment,
-                  redirect_uri: "http://redirect.uri"
+                  format: nil,
+                  redirect_uri: nil
                 }}
     end
 
@@ -90,8 +90,8 @@ defmodule Boruta.OauthTest.PreauthorizeTest do
                   error: :invalid_client,
                   error_description: "Invalid client_id or redirect_uri.",
                   status: :unauthorized,
-                  format: :fragment,
-                  redirect_uri: "http://bad.redirect.uri"
+                  format: nil,
+                  redirect_uri: nil
                 }}
     end